Remove scala.quoted.unsafe.UnsafeExpr
The library should not provide those operattion on Expr.
These can still be done using reflection and we keep testing that they can be encoded.
