+++ tests/run-macros/quote-matcher-symantics-2/quoted_1.scala
@@ -1,7 +1,5 @@
 import scala.quoted._
 
-import scala.quoted.unsafe._
-
 object Macros {
 
   inline def liftString(inline a: DSL): String = ${implStringNum('a)}
@@ -62,6 +60,29 @@ object Macros {
     lift(a)(Map.empty)
   }
 
+}
+
+object UnsafeExpr {
+  def open[T1, R, X](f: Expr[T1 => R])(content: (Expr[R], [t] => Expr[t] => Expr[T1] => Expr[t]) => X)(using qctx: QuoteContext): X = {
+    val (params, bodyExpr) = paramsAndBody[R](f)
+    content(bodyExpr, [t] => (e: Expr[t]) => (v: Expr[T1]) => bodyFn[t](e.unseal, params, List(v.unseal)).seal.asInstanceOf[Expr[t]])
+  }
+  private def paramsAndBody[R](using qctx: QuoteContext)(f: Expr[Any]): (List[qctx.reflect.ValDef], Expr[R]) = {
+    import qctx.reflect._
+    val Block(List(DefDef("$anonfun", Nil, List(params), _, Some(body))), Closure(Ident("$anonfun"), None)) = f.unseal.etaExpand
+    (params, body.seal.asInstanceOf[Expr[R]])
+  }
+
+  private def bodyFn[t](using qctx: QuoteContext)(e: qctx.reflect.Term, params: List[qctx.reflect.ValDef], args: List[qctx.reflect.Term]): qctx.reflect.Term = {
+    import qctx.reflect._
+    val map = params.map(_.symbol).zip(args).toMap
+    new TreeMap {
+      override def transformTerm(tree: Term)(using ctx: Context): Term =
+        super.transformTerm(tree) match
+          case tree: Ident => map.getOrElse(tree.symbol, tree)
+          case tree => tree
+    }.transformTerm(e)
+  }
 }
 
 def freshEnvVar()(using QuoteContext): (Int, Expr[DSL]) = {
